Move padding function to padding module
