+++ transformers4rec/torch/utils/data_utils.py
@@ -17,11 +17,9 @@
 import logging
 import warnings
 from abc import ABC
-from typing import Dict, Optional
 
 import numpy as np
 import torch
-import torch.nn.functional as F
 from merlin.dataloader.torch import Loader
 from merlin.models.utils.misc_utils import validate_dataset
 from merlin.models.utils.registry import Registry
@@ -30,6 +28,8 @@ from torch.utils.data import DataLoader as PyTorchDataLoader
 from torch.utils.data import Dataset, IterableDataset
 
 from merlin_standard_lib import Schema
+from transformers4rec.torch.utils import schema_utils
+from transformers4rec.torch.utils.padding import get_pad_fn
 
 from ...utils import dependencies
 
@@ -343,7 +343,9 @@ class MerlinDataLoader(T4RecDataLoader, DLDataLoader):
                 " for DDP training and ensure optimal performance."
             )
 
-        self.dataset.schema = self.dataset.schema.select_by_name(cats + conts + labels)
+        self.dataset.schema = schema_utils._augment_schema(
+            self.dataset.schema, cats=cats, conts=conts, labels=labels
+        )
 
         loader = Loader(
             self.dataset,
@@ -355,7 +357,7 @@ class MerlinDataLoader(T4RecDataLoader, DLDataLoader):
             global_size=global_size,
             global_rank=global_rank,
             drop_last=drop_last,
-        ).map(_get_pad_fn(sparse_max))
+        ).map(get_pad_fn(sparse_max))
 
         DLDataLoader.__init__(
             self,
@@ -437,91 +439,6 @@ class MerlinDataLoader(T4RecDataLoader, DLDataLoader):
         )
 
         return loader
-
-
-def _pad_dense_tensor(t: torch.Tensor, length: Optional[int]) -> torch.Tensor:
-    if length and len(t.shape) == 2:
-        pad_diff = length - t.shape[1]
-        return F.pad(input=t, pad=(0, pad_diff, 0, 0))
-    return t
-
-
-def _squeeze(tensor):
-    if len(tensor.shape) == 2:
-        return tensor.squeeze(1)
-    return tensor
-
-
-def _pad_ragged_tensor_1(values: torch.Tensor, offsets: torch.Tensor, padding_length: int):
-    values = _squeeze(values)
-    offsets = _squeeze(offsets)
-    num_rows = len(offsets) - 1
-    padded_values = torch.zeros(
-        (num_rows, padding_length), dtype=values.dtype, device=values.device
-    )
-    for i in range(num_rows):
-        row_values = values[offsets[i] : offsets[i + 1]]
-        padded_values[i, : len(row_values)] = row_values
-    return padded_values
-
-
-def _get_indices(offsets, diff_offsets):
-    row_ids = torch.arange(len(offsets) - 1, device=offsets.device)
-    row_ids_repeated = torch.repeat_interleave(row_ids, diff_offsets)
-    row_offset_repeated = torch.repeat_interleave(offsets[:-1], diff_offsets)
-    col_ids = torch.arange(len(row_offset_repeated), device=offsets.device) - row_offset_repeated
-    indices = torch.cat([row_ids_repeated.unsqueeze(-1), col_ids.unsqueeze(-1)], axis=1)
-    return indices
-
-
-def _pad_ragged_tensor_2(values, offsets, padding_length):
-    values = _squeeze(values)
-    offsets = _squeeze(offsets)
-    num_rows = len(offsets) - 1
-    diff_offsets = offsets[1:] - offsets[:-1]
-    indices = _get_indices(offsets, diff_offsets)
-    sparse_tensor = torch.sparse_coo_tensor(
-        indices.T, values, torch.Size([num_rows, padding_length]), device=values.device
-    )
-    return sparse_tensor.to_dense()
-
-
-def _pad_batch(X, padding_lengths, ragged_pad_fn):
-    if not X:
-        return X
-
-    X_padded = {}
-    for k, values in X.items():
-        if k.endswith("__values"):
-            col_name = k[:-8]
-            offsets = X[f"{col_name}__offsets"]
-            padding_length = padding_lengths.get(col_name)
-            padded_values = ragged_pad_fn(values, offsets, padding_length)
-            X_padded[col_name] = padded_values
-        elif k.endswith("__offsets"):
-            continue
-        elif isinstance(values, tuple):
-            padding_length = padding_lengths.get(k)
-            if padding_length:
-                values, offsets = values
-                padded_values = ragged_pad_fn(values, offsets, padding_length)
-                X_padded[k] = padded_values
-            else:
-                X_padded[k] = values
-        else:
-            padding_length = padding_lengths.get(k)
-            X_padded[k] = _pad_dense_tensor(values, padding_length)
-
-    return X_padded
-
-
-def _get_pad_fn(padding_lengths: Dict[str, int]):
-    def pad_fn(x, y):
-        new_x = _pad_batch(x, padding_lengths, _pad_ragged_tensor_2)
-        new_y = _pad_batch(y, padding_lengths, _pad_ragged_tensor_2)
-        return new_x, new_y
-
-    return pad_fn
 
 
 class ParquetDataset(Dataset):
